+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 
 * :bug:`3406` Sushiswap events will now be taken into account during the profit and loss report.
 * :bug:`3407` Users will now see correctly reported their amount of xSUSHI.
+* :bug:`3421` Users will now be able to see solana properly in coinbase assets
 
 * :release:`1.20.1 <2021-08-27>`
 * :bug:`3349` Defi swaps trade now shows in trade history.